server/requirements.txt
django>=4.0.0
djangorestframework
mysqlclient